// frontend/src/api/generated/models/worklogRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable issue tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */
import type { UserPublic } from './userPublic';

export interface WorklogRead {
  id: number;
  issue_id: number;
  user: UserPublic;
  minutes: number;
  worked_on: string;
  note?: string | null;
  created_at: string;
  updated_at: string;
}